test(cli-e2e): split studio init tests into interactive and non-interactive files
Improves E2E shard balance by splitting the dominant 848-line file into
two files (~537 + ~269 lines) so vitest can distribute them across shards.
